Refactor: Extract tilt optimization and mesh repair from Minimizer
