+++ runtime/minimizer.py
@@ -22,7 +22,6 @@ from runtime.diagnostics.audit import (
 from runtime.diagnostics.gauss_bonnet import GaussBonnetMonitor
 from runtime.energy_context import EnergyContext
 from runtime.energy_manager import EnergyModuleManager
-from runtime.equiangulation import equiangulate_iteration
 from runtime.evaluation_manager import EvaluationManager
 from runtime.interface_validation import validate_disk_interface_topology
 from runtime.leaflet_validation import validate_leaflet_absence_topology
@@ -37,6 +36,9 @@ from runtime.projections.tilt import (
 )
 from runtime.steppers.base import BaseStepper
 from runtime.steppers.tilt_relaxation import TiltRelaxationManager
+
+from .mesh_quality_repair import _maybe_auto_mesh_quality_repair
+from .tilt_optimization import _optimize_thetaB_scalar
 
 logger = logging.getLogger("membrane_solver")
 
@@ -937,147 +939,6 @@ STEP SIZE:\t {self.step_size}
         self.mesh.touch_tilts_in()
         self.mesh.touch_tilts_out()
 
-    def _optimize_thetaB_scalar(self, *, tilt_mode: str, iteration: int) -> None:
-        """Optionally optimize the scalar thetaB by sampling reduced energies.
-
-        This treats thetaB as a global scalar degree of freedom and updates it
-        by comparing the total energy after a partial tilt relaxation for a few
-        candidate thetaB values.
-        """
-        mode_match = (
-            str(self.global_params.get("rim_slope_match_mode") or "").strip().lower()
-        )
-        trace_radius = self.global_params.get("parity_trace_layer_radius")
-        outer_shells = int(self.global_params.get("parity_outer_shells", 0) or 0)
-        if (
-            mode_match == "physical_edge_staggered_v1"
-            and trace_radius is not None
-            and outer_shells > 0
-        ):
-            return
-        if not bool(self.global_params.get("tilt_thetaB_optimize", False)):
-            return
-
-        every = int(self.global_params.get("tilt_thetaB_optimize_every", 10) or 10)
-        if every <= 0:
-            every = 1
-        if int(iteration) % every != 0:
-            return
-
-        delta = float(self.global_params.get("tilt_thetaB_optimize_delta", 0.02) or 0.0)
-        if delta <= 0.0:
-            return
-
-        # Warm-start from the current relaxed state.
-        base_thetaB = float(self.global_params.get("tilt_thetaB_value") or 0.0)
-        base_tin = self.mesh.tilts_in_view().copy(order="F")
-        base_tout = self.mesh.tilts_out_view().copy(order="F")
-
-        # Use a smaller inner relaxation budget for the thetaB scan to keep the
-        # optimization cheap, but still responsive.
-        orig_inner_steps = self.global_params.get("tilt_inner_steps", None)
-        scan_steps = int(
-            self.global_params.get("tilt_thetaB_optimize_inner_steps", 20) or 20
-        )
-        if scan_steps < 1:
-            scan_steps = 1
-        self.global_params.set("tilt_inner_steps", scan_steps)
-
-        # Guard threshold prevents thetaB scan candidates from diverging.
-        scan_guard_factor = float(
-            self.global_params.get("tilt_relax_energy_guard_factor", 0.0) or 0.0
-        )
-
-        def eval_candidate(thetaB_val: float) -> tuple[float, np.ndarray, np.ndarray]:
-            self.global_params.set("tilt_thetaB_value", float(thetaB_val))
-            self._set_leaflet_tilts_from_arrays_fast(base_tin, base_tout)
-            # Relax tilts only; shape is handled by the main loop.
-            self._relax_leaflet_tilts(
-                positions=self.mesh.positions_view(), mode=tilt_mode
-            )
-            e = float(self.compute_energy())
-            # If the candidate's energy is much worse than the baseline,
-            # discard it to prevent tilt divergence from corrupting
-            # the selection.
-            if scan_guard_factor > 0.0:
-                scan_threshold = max(
-                    float(
-                        self.global_params.get("tilt_relax_energy_guard_min", 1e-4)
-                        or 1e-4
-                    ),
-                    abs(e0) * scan_guard_factor,
-                )
-                if e > scan_threshold:
-                    if logger.isEnabledFor(logging.DEBUG):
-                        logger.debug(
-                            "thetaB scan: thetaB=%.6g candidate E=%.6g "
-                            "exceeds guard threshold %.6g; discarding.",
-                            thetaB_val,
-                            e,
-                            scan_threshold,
-                        )
-                    # Return sentinel: restore base tilts and flag as bad.
-                    self._set_leaflet_tilts_from_arrays_fast(base_tin, base_tout)
-                    return (float("inf"), base_tin, base_tout)
-            return (
-                e,
-                self.mesh.tilts_in_view().copy(order="F"),
-                self.mesh.tilts_out_view().copy(order="F"),
-            )
-
-        try:
-            e0 = float(self.compute_energy())
-            e_minus, tin_minus, tout_minus = eval_candidate(base_thetaB - delta)
-            e_plus, tin_plus, tout_plus = eval_candidate(base_thetaB + delta)
-        finally:
-            # Restore the user's configured inner step budget.
-            if orig_inner_steps is None:
-                self.global_params.unset("tilt_inner_steps")
-            else:
-                self.global_params.set("tilt_inner_steps", orig_inner_steps)
-
-        # Pick the best among the sampled points (cheap coordinate descent).
-        best = min(
-            [
-                (e0, base_thetaB, base_tin, base_tout),
-                (e_minus, base_thetaB - delta, tin_minus, tout_minus),
-                (e_plus, base_thetaB + delta, tin_plus, tout_plus),
-            ],
-            key=lambda x: x[0],
-        )
-
-        best_e, best_thetaB, best_tin, best_tout = best
-        # Guard: if the best candidate is worse than the starting state,
-        # roll back to the original thetaB and tilts to prevent the
-        # optimizer from pushing the system into an unstable configuration
-        # (e.g. after mesh refinement when the tilt eigenvalue spectrum
-        # has changed).
-        if best_e > e0:
-            self.global_params.set("tilt_thetaB_value", float(base_thetaB))
-            self._set_leaflet_tilts_from_arrays_fast(base_tin, base_tout)
-            if logger.isEnabledFor(logging.WARNING):
-                logger.warning(
-                    "thetaB optimize: i=%d rollback (best E %.6g > base E %.6g); "
-                    "keeping thetaB=%.6g.",
-                    int(iteration),
-                    best_e,
-                    e0,
-                    base_thetaB,
-                )
-        else:
-            self.global_params.set("tilt_thetaB_value", float(best_thetaB))
-            self._set_leaflet_tilts_from_arrays_fast(best_tin, best_tout)
-
-        if logger.isEnabledFor(logging.DEBUG):
-            logger.debug(
-                "thetaB optimize: i=%d thetaB %.6g -> %.6g (E %.6g -> %.6g)",
-                int(iteration),
-                base_thetaB,
-                float(self.global_params.get("tilt_thetaB_value") or 0.0),
-                e0,
-                best_e,
-            )
-
     def enforce_constraints_after_mesh_ops(self, mesh: Mesh | None = None):
         """Enforce constraints after discrete mesh operations."""
         if not self._has_enforceable_constraints:
@@ -1100,66 +961,6 @@ STEP SIZE:\t {self.step_size}
         # mesh version so cached SoA views (positions/tilts/curvature) are not
         # stale after mesh operations like refinement/averaging.
         target_mesh.increment_version()
-
-    def _triangle_aspect_percentile(self, percentile: float = 90.0) -> float:
-        """Return triangle aspect-ratio percentile ``h_max/h_min`` for current mesh."""
-        tri_rows, _ = self.mesh.triangle_row_cache()
-        if tri_rows is None or len(tri_rows) == 0:
-            return float("nan")
-        pos = self.mesh.positions_view()
-        tri = pos[np.asarray(tri_rows, dtype=int)]
-        e01 = np.linalg.norm(tri[:, 0] - tri[:, 1], axis=1)
-        e12 = np.linalg.norm(tri[:, 1] - tri[:, 2], axis=1)
-        e20 = np.linalg.norm(tri[:, 2] - tri[:, 0], axis=1)
-        h_max = np.maximum.reduce([e01, e12, e20])
-        h_min = np.minimum.reduce([e01, e12, e20])
-        ratio = h_max / np.maximum(h_min, 1e-18)
-        return float(np.percentile(ratio, float(percentile)))
-
-    def _maybe_auto_mesh_quality_repair(self, *, iteration: int) -> bool:
-        """Optionally run bounded mesh-quality repair via equiangulation passes."""
-        if not bool(self.global_params.get("mesh_quality_auto_repair_enabled", False)):
-            return False
-        every = int(self.global_params.get("mesh_quality_auto_repair_every", 0) or 0)
-        if every <= 0 or ((int(iteration) + 1) % every) != 0:
-            return False
-
-        threshold = float(
-            self.global_params.get("mesh_quality_aspect_threshold", 0.0) or 0.0
-        )
-        if threshold <= 0.0:
-            return False
-        perc = float(
-            self.global_params.get("mesh_quality_aspect_percentile", 90.0) or 90.0
-        )
-        max_passes = int(
-            self.global_params.get("mesh_quality_max_repair_passes", 1) or 1
-        )
-        if max_passes <= 0:
-            return False
-
-        aspect_p = self._triangle_aspect_percentile(percentile=perc)
-        if not np.isfinite(aspect_p) or aspect_p <= threshold:
-            return False
-
-        changed_any = False
-        for _ in range(max_passes):
-            new_mesh, changed = equiangulate_iteration(self.mesh)
-            if not bool(changed):
-                break
-            self.mesh = new_mesh
-            self.enforce_constraints_after_mesh_ops(self.mesh)
-            self.mesh.project_tilts_to_tangent()
-            self.mesh.increment_version()
-            changed_any = True
-            aspect_p = self._triangle_aspect_percentile(percentile=perc)
-            if not np.isfinite(aspect_p) or aspect_p <= threshold:
-                break
-        if changed_any:
-            reset = getattr(self.stepper, "reset", None)
-            if callable(reset):
-                reset()
-        return changed_any
 
     def minimize(
         self, n_steps: int = 1, callback: Optional[Callable[["Mesh", int], None]] = None
@@ -1273,7 +1074,7 @@ STEP SIZE:\t {self.step_size}
 
             self._update_scalar_params()
             if self._uses_leaflet_tilts():
-                self._optimize_thetaB_scalar(tilt_mode=str(tilt_mode), iteration=i)
+                _optimize_thetaB_scalar(self, tilt_mode=str(tilt_mode), iteration=i)
 
             # Use array-based path for main loop
             E, grad_arr = self.compute_energy_and_gradient_array()
@@ -1456,7 +1257,7 @@ STEP SIZE:\t {self.step_size}
                         if callable(reset):
                             reset()
 
-                self._maybe_auto_mesh_quality_repair(iteration=i)
+                _maybe_auto_mesh_quality_repair(self, iteration=i)
 
         if self._has_enforceable_constraints:
             # One final projection improves cross-platform determinism for hard
